backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
aiohttp
pandas
sodapy